Certificate Generator/requirements.txt
pandas==1.4.2
python_pptx==0.6.21
pywin32==304
